Changed getValueFromSource to return Promise
Serverless from version 1.16.0 needs getValueFromSource to return
Promise (before it was alredy resolved value). Introduced changes to
support that. Additionaly upgraded ava and enforced serial execution on
tests which create and clean directories as they happened to fail when
run in parallel. Upgraded serverless as well and tested manually with
version 1.17.0.