Better value for git:branch when in detached state
Certain CI services may deploy a git repo by checking out a specific
commit rather than the branch. When this happens, the repo is in a
"detached" state. Currently, this causes the value for `git:branch` to
always be HEAD instead of the name of the branch associated with the
commit.

If the commit happens to be the last commit of the branch, it is possible
to retrieve the branch name and use that instead of HEAD.

This commit makes a minor change to the `branch` case in `_getValueFromGit`
so that if the initial value is HEAD it attempts to determine the correct
branch name, falling back on HEAD if it is not able to.

This commit includes 2 test cases to ensure correct behavior along with
a test repo `multiple_commits` used for testing.